refactor(server): Combine mcpClientFactory into mcpManager
